requirements.txt
# asgiref==3.3.4
# certifi==2020.12.5
# cffi==1.14.5
# chardet==4.0.0
# coreapi==2.3.3
# coreschema==0.0.4
# cryptography==3.3.1
# Django==2.2.3
# django-cors-headers==3.7.0
# django-oauth-toolkit==1.2.0
# django-rest-framework==0.1.0
# django-simple-search==1.0.2
# djangorestframework==3.9.4
# drf-yasg==1.16.0
# idna==2.10
# inflection==0.5.1
# itypes==1.2.0
# Jinja2==3.0.1
# MarkupSafe==2.0.1
# mysqlclient==1.4.2.post1
# oauthlib==3.1.0
# pkg-resources==0.0.0
# pycparser==2.20
# pytz==2021.1
# requests==2.25.1
# ruamel.yaml==0.17.4
# ruamel.yaml.clib==0.2.2
# six==1.16.0
# sqlparse==0.4.1
# typing-extensions==3.10.0.0
# uritemplate==3.0.1
# urllib3==1.26.4




chardet==5.0.0
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
Deprecated==1.2.13
Django==4.1.7
django-cors-headers==3.14.0
django-oauth-toolkit==2.2.0
django-oauth2-provider==0.2.6.1
django-simple-search==1.0.2
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.7.1
requests==2.28.2
pkg-resources==0.0.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
shortuuid==1.0.11
mysqlclient==1.4.2.post1
six==1.16.0
sqlparse==0.4.3
typing-extensions==3.10.0.0
stark-utilities==0.0.5
uritemplate==4.1.1
urllib3==1.26.15
wrapt==1.15.0